@@ -5,6 +5,12 @@ namespace CompanyName.MyMeetings.Modules.Meetings.Application.Configuration.Comm
 {
     public interface ICommandHandler<in TCommand> : IRequestHandler<TCommand> where TCommand : ICommand
     {
-        
+
+    }
+
+    public interface ICommandHandler<in TCommand, TResult> :
+        IRequestHandler<TCommand, TResult> where TCommand : ICommand<TResult>
+    {
+
     }
 }
