@@ -2,11 +2,26 @@
 
 namespace CompanyName.MyMeetings.Modules.Meetings.Application.Contracts
 {
-    public class CommandBase : ICommand
+    public abstract class CommandBase : ICommand
     {
         public Guid Id { get; }
 
-        public CommandBase()
+        protected CommandBase()
+        {
+            this.Id = Guid.NewGuid();
+        }
+
+        protected CommandBase(Guid id)
+        {
+            this.Id = id;
+        }
+    }
+
+    public abstract class CommandBase<TResult> : ICommand<TResult>
+    {
+        public Guid Id { get; }
+
+        protected CommandBase()
         {
             this.Id = Guid.NewGuid();
         }
